# api/docs/swagger.yaml
basePath: /auth
definitions:
  models.AccessToken:
    properties:
      access_token:
        type: string
    type: object
  models.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.RefreshToken:
    properties:
      refresh_token:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role:
        type: string
    type: object
  models.RegisterResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
    type: object
  models.Tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Authorazation API of On-Demand Car Wash Service
  title: Authorazation
  version: "1.0"
paths:
  /login:
    post:
      description: Logs user in
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      summary: Login user
      tags:
      - auth
  /logout:
    post:
      description: Logouts user
      parameters:
      - description: User email
        in: query
        name: email
        required: true
        type: string
      responses:
        "200":
          description: User logged out successfully
          schema:
            type: string
        "400":
          description: Invalid email
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      summary: Logouts user
      tags:
      - auth
  /refresh:
    post:
      description: Refreshes refresh token
      parameters:
      - description: Refresh token
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.RefreshToken'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      summary: Refreshes token
      tags:
      - auth
  /register:
    post:
      description: Registers a new user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      summary: Registers user
      tags:
      - auth
  /validate:
    post:
      description: Validates access token
      parameters:
      - description: Access token
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.AccessToken'
      responses:
        "200":
          description: Access token is valid
          schema:
            type: string
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      summary: Validates token
      tags:
      - auth
swagger: "2.0"